changes to styling to keep banners for shared ideas, login, signup, and your ideas to be fixed under nav bar, moved creat new idea button and position based on screen size

+++ src/components/IdeaCard.js
@@ -11,7 +11,10 @@ class IdeaCard extends Component {
     }
 
     handleDelete = () => {
-        this.props.deleteIdea(this.props.ideas)
+        let confirmBox = window.confirm("Are you sure you want to delete this idea?")
+            if (confirmBox === true){
+                this.props.deleteIdea(this.props.ideas)
+            }
     }
 
     handlePrivate = () => {
